Update index.js, int.js
