+++ src/index.html
@@ -5,24 +5,24 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>To-Do List</title>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" integrity="sha512-iecdLmaskl7CVkqkXNQ/ZH/XLlvWZOJyj7Yy7tcenmpD1ypASozpmT/E0iPtmFIB46ZmdtAc9eNBvH0H/ZpiBw==" crossorigin="anonymous" referrerpolicy="no-referrer">
-
-  </head>
+    <link rel="stylesheet" href="style.css">
+</head>
 <body>
     <section class="card">
         <div class="head">
-            <p>Today's To Do</p>
+            <p>To Do List</p>
             <i class="fa-solid fa-rotate rotate"></i>
         </div>
-  
+        <hr>
         <div class="addData">
             <input id="task" type="text" placeholder="Add to your List...." required>
             <button type="button"> <i id="add-button" class="fa-solid fa-turn-down turnLeft"></i> </button>
         </div>
-    
+        <hr>
         <ul id="lists" class="lists">
                 
         </ul>
-        <button class="clear-btn" type="button">Clear all completed</button>
+        <button id="clearCompletedBtn" class="clear-btn" type="button">Clear all completed</button>
     </section>
 
     <script type="module" src="index.js"></script>
