+++ src/styles/style.css
@@ -1,7 +1,14 @@
+@import url('https://fonts.googleapis.com/css2?family=Fuzzy+Bubbles:wght@700&display=swap');
+
 * {
-  margin: 5px;
-  padding: 10px;
-  font-size: 24px;
+  margin: 0;
+  padding: 0;
+  font-family: 'Fuzzy Bubbles', cursive;
+  font-size: 20px;
+}
+
+body {
+  background-color: #eee;
 }
 
 .card {
@@ -11,7 +18,6 @@
   margin: 10rem auto;
   padding: 20px;
   background-color: #fff;
-  border: 1px solid #000;
 }
 
 .head {
@@ -36,6 +42,7 @@
 
 input {
   border: none;
+  outline: none;
 }
 
 input[type="text"]:focus {
